+++ mpsl/include/mpsl_fem_config_common.h
@@ -34,6 +34,15 @@ typedef struct
     uint8_t gpiote_ch_id; /**< The GPIOTE channel used for toggling this pin. */
 } mpsl_fem_gpiote_pin_config_t;
 
+/** @brief Configuration parameters for pins that enable or disable (or both) either Power Amplifier (PA) or Low Noise Amplifier (LNA).
+ */
+typedef struct
+{
+    bool    enable;       /**< Enable toggling for this pin. */
+    bool    active_high;  /**< If true, the pin will be active high. Otherwise, the pin will be active low. */
+    uint8_t gpio_pin;     /**< GPIO pin number for the pin. */
+} mpsl_fem_gpio_pin_config_t;
+
 /** @brief Sets flag which determines whether DEVICE-CONFIG-254 shall be applied.
  *
  * @note This function shall be called before initialization of any protocol stack.